[C-IRIS] split cspace_free_polytope_test. (#18632)
