+++ geometry/optimization/dev/test/c_iris_test_utilities.h
@@ -1,10 +1,15 @@
 #pragma once
 
+#include <array>
 #include <memory>
+#include <optional>
+#include <unordered_map>
+#include <unordered_set>
 #include <vector>
 
 #include <gtest/gtest.h>
 
+#include "drake/geometry/optimization/dev/cspace_free_polytope.h"
 #include "drake/geometry/scene_graph.h"
 #include "drake/multibody/plant/multibody_plant.h"
 #include "drake/systems/framework/diagram.h"
@@ -53,6 +58,164 @@ class CIrisRobotPolytopicGeometryTest : public ::testing::Test {
   std::vector<geometry::GeometryId> world_boxes_;
   geometry::GeometryId world_convex_;
   std::vector<geometry::GeometryId> body_boxes_;
+};
+
+// This is a friend class of CspaceFreePolytope, we use it to expose the private
+// functions in CspaceFreePolytope for unit testing.
+class CspaceFreePolytopeTester {
+ public:
+  DRAKE_NO_COPY_NO_MOVE_NO_ASSIGN(CspaceFreePolytopeTester)
+
+  CspaceFreePolytopeTester(const multibody::MultibodyPlant<double>* plant,
+                           const geometry::SceneGraph<double>* scene_graph,
+                           SeparatingPlaneOrder plane_order,
+                           const Eigen::Ref<const Eigen::VectorXd>& q_star,
+                           const CspaceFreePolytope::Options& options =
+                               CspaceFreePolytope::Options())
+      : cspace_free_polytope_{new CspaceFreePolytope(
+            plant, scene_graph, plane_order, q_star, options)} {}
+
+  const CspaceFreePolytope& cspace_free_polytope() const {
+    return *cspace_free_polytope_;
+  }
+
+  void FindRedundantInequalities(
+      const Eigen::MatrixXd& C, const Eigen::VectorXd& d,
+      const Eigen::VectorXd& s_lower, const Eigen::VectorXd& s_upper,
+      double tighten, std::unordered_set<int>* C_redundant_indices,
+      std::unordered_set<int>* s_lower_redundant_indices,
+      std::unordered_set<int>* s_upper_redundant_indices) const {
+    cspace_free_polytope_->FindRedundantInequalities(
+        C, d, s_lower, s_upper, tighten, C_redundant_indices,
+        s_lower_redundant_indices, s_upper_redundant_indices);
+  }
+
+  const Eigen::VectorXd& s_lower() const {
+    return cspace_free_polytope_->s_lower_;
+  }
+
+  const Eigen::VectorXd& s_upper() const {
+    return cspace_free_polytope_->s_upper_;
+  }
+
+  const VectorX<symbolic::Polynomial>& s_minus_s_lower() const {
+    return cspace_free_polytope_->s_minus_s_lower_;
+  }
+
+  const VectorX<symbolic::Polynomial>& s_upper_minus_s() const {
+    return cspace_free_polytope_->s_upper_minus_s_;
+  }
+
+  const std::vector<PlaneSeparatesGeometries>& plane_geometries() const {
+    return cspace_free_polytope_->plane_geometries_;
+  }
+
+  template <typename T>
+  [[nodiscard]] VectorX<symbolic::Polynomial> CalcDminusCs(
+      const Eigen::Ref<const MatrixX<T>>& C,
+      const Eigen::Ref<const VectorX<T>>& d) const {
+    return cspace_free_polytope_->CalcDminusCs<T>(C, d);
+  }
+
+  [[nodiscard]] const std::unordered_map<
+      SortedPair<multibody::BodyIndex>,
+      std::array<VectorX<symbolic::Monomial>, 4>>&
+  map_body_to_monomial_basis_array() const {
+    return cspace_free_polytope_->map_body_to_monomial_basis_array_;
+  }
+
+  [[nodiscard]] CspaceFreePolytope::SeparationCertificateProgram
+  ConstructPlaneSearchProgram(
+      const PlaneSeparatesGeometries& plane_geometries,
+      const VectorX<symbolic::Polynomial>& d_minus_Cs,
+      const std::unordered_set<int>& C_redundant_indices,
+      const std::unordered_set<int>& s_lower_redundant_indices,
+      const std::unordered_set<int>& s_upper_redundant_indices) const {
+    return cspace_free_polytope_->ConstructPlaneSearchProgram(
+        plane_geometries, d_minus_Cs, C_redundant_indices,
+        s_lower_redundant_indices, s_upper_redundant_indices);
+  }
+
+  [[nodiscard]] std::vector<
+      std::optional<CspaceFreePolytope::SeparationCertificateResult>>
+  FindSeparationCertificateGivenPolytope(
+      const CspaceFreePolytope::IgnoredCollisionPairs& ignored_collision_pairs,
+      const Eigen::Ref<const Eigen::MatrixXd>& C,
+      const Eigen::Ref<const Eigen::VectorXd>& d,
+      const CspaceFreePolytope::FindSeparationCertificateGivenPolytopeOptions&
+          options) const {
+    return cspace_free_polytope_->FindSeparationCertificateGivenPolytope(
+        ignored_collision_pairs, C, d, options);
+  }
+
+  [[nodiscard]] int GetGramVarSizeForPolytopeSearchProgram(
+      const CspaceFreePolytope::IgnoredCollisionPairs& ignored_collision_pairs,
+      bool search_s_bounds_lagrangians) const {
+    return cspace_free_polytope_->GetGramVarSizeForPolytopeSearchProgram(
+        ignored_collision_pairs, search_s_bounds_lagrangians);
+  }
+
+  [[nodiscard]] std::unique_ptr<solvers::MathematicalProgram>
+  InitializePolytopeSearchProgram(
+      const CspaceFreePolytope::IgnoredCollisionPairs& ignored_collision_pairs,
+      const MatrixX<symbolic::Variable>& C,
+      const VectorX<symbolic::Variable>& d,
+      const VectorX<symbolic::Polynomial> d_minus_Cs,
+      const std::vector<
+          std::optional<CspaceFreePolytope::SeparationCertificateResult>>&
+          certificates_vec,
+      bool search_s_bounds_lagrangians, int gram_total_size,
+      std::unordered_map<int, CspaceFreePolytope::SeparationCertificate>*
+          new_certificates) const {
+    return cspace_free_polytope_->InitializePolytopeSearchProgram(
+        ignored_collision_pairs, C, d, d_minus_Cs, certificates_vec,
+        search_s_bounds_lagrangians, gram_total_size, new_certificates);
+  }
+
+  void AddEllipsoidContainmentConstraint(
+      solvers::MathematicalProgram* prog, const Eigen::MatrixXd& Q,
+      const Eigen::VectorXd& s0, const MatrixX<symbolic::Variable>& C,
+      const VectorX<symbolic::Variable>& d,
+      const VectorX<symbolic::Variable>& ellipsoid_margins) const {
+    cspace_free_polytope_->AddEllipsoidContainmentConstraint(prog, Q, s0, C, d,
+                                                             ellipsoid_margins);
+  }
+
+  void AddCspacePolytopeContainment(solvers::MathematicalProgram* prog,
+                                    const MatrixX<symbolic::Variable>& C,
+                                    const VectorX<symbolic::Variable>& d,
+                                    const Eigen::MatrixXd& s_inner_pts) const {
+    cspace_free_polytope_->AddCspacePolytopeContainment(prog, C, d,
+                                                        s_inner_pts);
+  }
+
+  [[nodiscard]] std::optional<
+      CspaceFreePolytope::FindPolytopeGivenLagrangianResult>
+  FindPolytopeGivenLagrangian(
+      const CspaceFreePolytope::IgnoredCollisionPairs& ignored_collision_pairs,
+      const MatrixX<symbolic::Variable>& C,
+      const VectorX<symbolic::Variable>& d,
+      const VectorX<symbolic::Polynomial>& d_minus_Cs,
+      const std::vector<
+          std::optional<CspaceFreePolytope::SeparationCertificateResult>>&
+          certificates_vec,
+      const Eigen::MatrixXd& Q, const Eigen::VectorXd& s0,
+      const VectorX<symbolic::Variable>& ellipsoid_margins, int gram_total_size,
+      const CspaceFreePolytope::FindPolytopeGivenLagrangianOptions& options,
+      std::unordered_map<int, CspaceFreePolytope::SeparationCertificateResult>*
+          certificates_result) const {
+    return cspace_free_polytope_->FindPolytopeGivenLagrangian(
+        ignored_collision_pairs, C, d, d_minus_Cs, certificates_vec, Q, s0,
+        ellipsoid_margins, gram_total_size, options, certificates_result);
+  }
+
+  HPolyhedron GetPolyhedronWithJointLimits(const Eigen::MatrixXd& C,
+                                           const Eigen::VectorXd& d) const {
+    return cspace_free_polytope_->GetPolyhedronWithJointLimits(C, d);
+  }
+
+ private:
+  std::unique_ptr<CspaceFreePolytope> cspace_free_polytope_;
 };
 
 }  // namespace optimization
